rename and move lib/conemesh.js to create_mesh.js
